Move functions to stackChunkOopDesc
