+++ src/hotspot/share/gc/shared/continuationGCSupport.inline.hpp
@@ -37,7 +37,7 @@ inline void ContinuationGCSupport::relativize_stack_chunk(oop obj) {
   }
   stackChunkOop chunk = stackChunkOopDesc::cast(obj);
   if (!chunk->is_gc_mode()) {
-    InstanceStackChunkKlass::relativize_chunk(chunk);
+    chunk->relativize();
   }
 }
 
@@ -48,7 +48,7 @@ inline void ContinuationGCSupport::transform_stack_chunk(oop obj) {
   stackChunkOop chunk = stackChunkOopDesc::cast(obj);
   if (!chunk->is_gc_mode()) {
     // Relativize *and* transform to use an oop bitmap
-    InstanceStackChunkKlass::transform_chunk(chunk);
+    chunk->transform();
   }
 }
