+++ src/hotspot/share/oops/instanceStackChunkKlass.inline.hpp
@@ -67,16 +67,6 @@ inline size_t InstanceStackChunkKlass::bitmap_size(size_t stack_size_in_words) {
   return align_up(size_in_bits, BitsPerWord) >> LogBitsPerWord;
 }
 
-template <InstanceStackChunkKlass::barrier_type barrier, chunk_frames frame_kind, typename RegisterMapT>
-void InstanceStackChunkKlass::do_barriers(stackChunkOop chunk, const StackChunkFrameStream<frame_kind>& f,
-                                          const RegisterMapT* map) {
-  if (frame_kind == chunk_frames::MIXED) {
-    // we could freeze deopted frames in slow mode.
-    f.handle_deopted();
-  }
-  do_barriers0<barrier>(chunk, f, map);
-}
-
 template <typename T, class OopClosureType>
 void InstanceStackChunkKlass::oop_oop_iterate(oop obj, OopClosureType* closure) {
   stackChunkOop chunk = stackChunkOopDesc::cast(obj);
@@ -187,44 +177,6 @@ void InstanceStackChunkKlass::oop_oop_iterate_stack_helper(stackChunkOop chunk, 
       StackChunkOopIterateBitmapClosure<oop, OopClosureType> bitmap_closure(chunk, closure);
       chunk->bitmap().iterate(&bitmap_closure, chunk->bit_index_for((oop*)start), chunk->bit_index_for((oop*)end));
     }
-  }
-}
-
-template <class StackChunkFrameClosureType>
-inline void InstanceStackChunkKlass::iterate_stack(stackChunkOop obj, StackChunkFrameClosureType* closure) {
-  obj->has_mixed_frames() ? iterate_stack<chunk_frames::MIXED>(obj, closure)
-                          : iterate_stack<chunk_frames::COMPILED_ONLY>(obj, closure);
-}
-
-template <chunk_frames frame_kind, class StackChunkFrameClosureType>
-inline void InstanceStackChunkKlass::iterate_stack(stackChunkOop obj, StackChunkFrameClosureType* closure) {
-  const SmallRegisterMap* map = SmallRegisterMap::instance;
-  assert(!map->in_cont(), "");
-
-  StackChunkFrameStream<frame_kind> f(obj);
-  bool should_continue = true;
-
-  if (f.is_stub()) {
-    RegisterMap full_map((JavaThread*)nullptr, true, false, true);
-    full_map.set_include_argument_oops(false);
-
-    f.next(&full_map);
-
-    assert(!f.is_done(), "");
-    assert(f.is_compiled(), "");
-
-    should_continue = closure->do_frame(f, &full_map);
-    f.next(map);
-    f.handle_deopted(); // the stub caller might be deoptimized (as it's not at a call)
-  }
-  assert(!f.is_stub(), "");
-
-  for(; should_continue && !f.is_done(); f.next(map)) {
-    if (frame_kind == chunk_frames::MIXED) {
-      // in slow mode we might freeze deoptimized frames
-      f.handle_deopted();
-    }
-    should_continue = closure->do_frame(f, map);
   }
 }
