+++ src/hotspot/share/oops/stackChunkOop.inline.hpp
@@ -34,6 +34,7 @@
 #include "runtime/frame.inline.hpp"
 #include "runtime/handles.inline.hpp"
 #include "runtime/registerMap.hpp"
+#include "runtime/smallRegisterMap.inline.hpp"
 #include "utilities/macros.hpp"
 #include CPU_HEADER_INLINE(stackChunkOop)
 
@@ -150,6 +151,53 @@ inline bool stackChunkOopDesc::has_thaw_slowpath_condition() const { return flag
 
 inline bool stackChunkOopDesc::requires_barriers() {
   return Universe::heap()->requires_barriers(this);
+}
+
+template <stackChunkOopDesc::barrier_type barrier, chunk_frames frame_kind, typename RegisterMapT>
+void stackChunkOopDesc::do_barriers(const StackChunkFrameStream<frame_kind>& f, const RegisterMapT* map) {
+  if (frame_kind == chunk_frames::MIXED) {
+    // we could freeze deopted frames in slow mode.
+    f.handle_deopted();
+  }
+  do_barriers0<barrier>(f, map);
+}
+
+template <class StackChunkFrameClosureType>
+inline void stackChunkOopDesc::iterate_stack(StackChunkFrameClosureType* closure) {
+  has_mixed_frames() ? iterate_stack<chunk_frames::MIXED>(closure)
+                     : iterate_stack<chunk_frames::COMPILED_ONLY>(closure);
+}
+
+template <chunk_frames frame_kind, class StackChunkFrameClosureType>
+inline void stackChunkOopDesc::iterate_stack(StackChunkFrameClosureType* closure) {
+  const SmallRegisterMap* map = SmallRegisterMap::instance;
+  assert(!map->in_cont(), "");
+
+  StackChunkFrameStream<frame_kind> f(this);
+  bool should_continue = true;
+
+  if (f.is_stub()) {
+    RegisterMap full_map((JavaThread*)nullptr, true, false, true);
+    full_map.set_include_argument_oops(false);
+
+    f.next(&full_map);
+
+    assert(!f.is_done(), "");
+    assert(f.is_compiled(), "");
+
+    should_continue = closure->do_frame(f, &full_map);
+    f.next(map);
+    f.handle_deopted(); // the stub caller might be deoptimized (as it's not at a call)
+  }
+  assert(!f.is_stub(), "");
+
+  for(; should_continue && !f.is_done(); f.next(map)) {
+    if (frame_kind == chunk_frames::MIXED) {
+      // in slow mode we might freeze deoptimized frames
+      f.handle_deopted();
+    }
+    should_continue = closure->do_frame(f, map);
+  }
 }
 
 inline frame stackChunkOopDesc::relativize(frame fr)   const { relativize_frame(fr);   return fr; }
